Se agrego el archivo de test que estaba ausente y se modifico el form desde el cual se borra la cursada
